@@ -17,8 +17,9 @@
     "@types/react-dom": "^17.0.13",
     "@usedapp/core": "^0.11.0",
     "framer-motion": "^6.2.8",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
